+++ empire/test/test_plugin_service.py
@@ -58,7 +58,6 @@ def test_auto_execute_plugins(caplog, monkeypatch, models, empire_config, instal
     caplog.set_level(logging.DEBUG)
 
     main_menu_mock = MagicMock()
-    main_menu_mock.installPath = str(install_path)
     main_menu_mock.install_path = Path(install_path)
     plugin_service = PluginService(main_menu_mock)
     plugin_service.startup()
@@ -72,7 +71,6 @@ def test_plugin_execute_with_kwargs(session_local, install_path):
         return f"This function was called with options: {options} and kwargs: {kwargs}"
 
     main_menu_mock = MagicMock()
-    main_menu_mock.installPath = install_path
     main_menu_mock.install_path = Path(install_path)
     plugin_service = PluginService(main_menu_mock)
     plugin_service.startup()
@@ -89,7 +87,6 @@ def test_plugin_execute_with_kwargs(session_local, install_path):
 
 def test_execute_plugin_file_option_not_found(install_path, session_local):
     main_menu_mock = MagicMock()
-    main_menu_mock.installPath = install_path
     main_menu_mock.install_path = Path(install_path)
 
     main_menu_mock.downloadsv2 = MagicMock()
@@ -124,7 +121,6 @@ def test_execute_plugin_file_option_not_found(install_path, session_local):
 
 def test_execute_plugin_file_option(install_path, session_local, models):
     main_menu_mock = MagicMock()
-    main_menu_mock.installPath = install_path
     main_menu_mock.install_path = Path(install_path)
 
     download = models.Download(id=9999, filename="test_file", location="/tmp/test_file")
@@ -174,7 +170,6 @@ def test_execute_plugin_file_option(install_path, session_local, models):
 # then it will fail.
 def test_on_start_on_stop_called(install_path):
     main_menu_mock = MagicMock()
-    main_menu_mock.installPath = install_path
     main_menu_mock.install_path = Path(install_path)
 
     plugin_service = PluginService(main_menu_mock)
@@ -199,7 +194,6 @@ def test_on_start_on_stop_called(install_path):
 # then it will fail.
 def test_on_load_on_unload_called(install_path):
     main_menu_mock = MagicMock()
-    main_menu_mock.installPath = install_path
     main_menu_mock.install_path = Path(install_path)
 
     plugin_service = PluginService(main_menu_mock)
@@ -274,7 +268,6 @@ def test__determine_auto_execute(empire_config, plugin_service):
 
 def test_plugin_load_exception(install_path, session_local):
     main_menu_mock = MagicMock()
-    main_menu_mock.installPath = install_path
     main_menu_mock.install_path = Path(install_path)
 
     plugin_service = PluginService(main_menu_mock)